[python-fronted] move string module files to string directory
